feat: add service and controller for purchase report

+++ src/main/java/org/shop/beautyportal/purchasereport/domain/entities/PurchaseReport.java
@@ -54,7 +54,7 @@ public class PurchaseReport {
             example = "2"
     )
     @Column(name = "quarter")
-    private Short quarter;
+    private Integer quarter;
 
     @Schema(
             description = "Sales value from the last year",
